% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arspi_functions.R
\name{arspi_estimate}
\alias{arspi_estimate}
\title{Compute ARSPI and Drought Characteristics}
\usage{
arspi_estimate(rainfall, scale, model_file = NULL)
}
\arguments{
\item{rainfall}{A numeric vector of monthly rainfall values.}

\item{scale}{An integer indicating the scale (e.g., 3, 6, 12).}
}
\value{
A list containing ARSPI values and drought characteristics for each drought category.
}
\description{
Compute ARSPI and Drought Characteristics
}
